+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 *   shrink `DigestClient`
 *   support `userhash` in `DigestClient`
+*   support `-sess` algorithm variants in `DigestClient`
 
 ## `v0.1.3` (2021-10-20)